refactor: Moved `website` directory out of `src`
