+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "./tsconfig.base",
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "website/**/*"],
   "compilerOptions": {
-    "module": "ES6"
+    "module": "ES6",
+    "outDir": "./dist"
   }
 }
