+++ website/components/AppWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement, useMemo, useState } from "react";
 import { useLocation, useSearchParams } from "react-router-dom";
-import { GlobalViewerSettings } from "../../aics-image-viewer/components/App/types";
-import { ImageViewerApp, RenderMode, ViewMode } from "../..";
+import { GlobalViewerSettings } from "../../src/aics-image-viewer/components/App/types";
+import { ImageViewerApp, RenderMode, ViewMode } from "../../src";
 import { getArgsFromParams } from "../utils/url_utils";
 import { AppDataProps } from "../types";
