+++ website/types.ts
@@ -1,4 +1,4 @@
-import { AppProps } from "../aics-image-viewer/components/App/types";
+import { AppProps } from "../src/aics-image-viewer/components/App/types";
 
 export type AppDataProps = Omit<AppProps, "appHeight" | "canvasMargin">;
