+++ website/utils/url_utils.ts
@@ -1,8 +1,8 @@
-import FirebaseRequest, { DatasetMetaData } from "../../../public/firebase";
+import FirebaseRequest, { DatasetMetaData } from "../../public/firebase";
 
-import { AppProps, GlobalViewerSettings } from "../../aics-image-viewer/components/App/types";
-import { ViewMode } from "../../aics-image-viewer/shared/enums";
-import { ViewerChannelSettings } from "../../aics-image-viewer/shared/utils/viewerChannelSettings";
+import { AppProps, GlobalViewerSettings } from "../../src/aics-image-viewer/components/App/types";
+import { ViewMode } from "../../src/aics-image-viewer/shared/enums";
+import { ViewerChannelSettings } from "../../src/aics-image-viewer/shared/utils/viewerChannelSettings";
 
 const paramKeys = ["mask", "ch", "luts", "colors", "url", "file", "dataset", "id", "view"];
 type ParamKeys = (typeof paramKeys)[number];
